TASK MANAGEMENT SYSTEM
In this web application i have used spring boot as the backend framework
which uses java language to create a server.I have used maven project
to configure and create my spring boot application.

And in the frontend i have used react framework and integerate it with
my backend.

IDE's used are intellij IDEA for backend and VSCode for frontend.